iterable for str をファイル分割
